moved getRollerRootDir to a higher level directory
