+++ cmd/config/init/error_handler.go
@@ -1,19 +1,1 @@
 package initconfig
-
-import (
-	"os"
-
-	"github.com/fatih/color"
-)
-
-func OutputCleanError(err error) {
-	if err != nil {
-		defer func() {
-			if r := recover(); r != nil {
-				os.Exit(1)
-			}
-		}()
-		color.New(color.FgRed, color.Bold).Printf("💈 %s\n", err.Error())
-		panic(err)
-	}
-}
