+++ cmd/consts/consts.go
@@ -7,7 +7,7 @@ var Executables = struct {
 	Dymension  string
 }{
 	Celestia:   "/usr/local/bin/roller_bins/celestia",
-	RollappEVM: "/usr/local/bin/roller_bins/rollapp_evm",
+	RollappEVM: "/usr/local/bin/rollapp_evm",
 	Relayer:    "/usr/local/bin/roller_bins/rly",
 	Dymension:  "/usr/local/bin/roller_bins/dymd",
 }
